Refactor: move autorotate logic from objects_along_edge to geom.py.
